+++ src/SAWScript/Crucible/JVM/ResolveSetupValue.hs
@@ -141,6 +141,7 @@ typeOfSetupValue _cc env _nameEnv val =
     MS.SetupGlobal empty _            -> absurd empty
     MS.SetupStruct empty _            -> absurd empty
     MS.SetupTuple empty _             -> absurd empty
+    MS.SetupSlice empty               -> absurd empty
     MS.SetupArray empty _             -> absurd empty
     MS.SetupElem empty _ _            -> absurd empty
     MS.SetupField empty _ _           -> absurd empty
@@ -173,6 +174,7 @@ resolveSetupVal cc env _tyenv _nameEnv val =
     MS.SetupGlobal empty _            -> absurd empty
     MS.SetupStruct empty _            -> absurd empty
     MS.SetupTuple empty _             -> absurd empty
+    MS.SetupSlice empty               -> absurd empty
     MS.SetupArray empty _             -> absurd empty
     MS.SetupElem empty _ _            -> absurd empty
     MS.SetupField empty _ _           -> absurd empty