+++ src/SAWScript/Crucible/JVM/Setup/Value.hs
@@ -79,6 +79,7 @@ type instance MS.XSetupNull CJ.JVM = ()
 type instance MS.XSetupGlobal CJ.JVM = Void
 type instance MS.XSetupStruct CJ.JVM = Void
 type instance MS.XSetupTuple CJ.JVM = Void
+type instance MS.XSetupSlice CJ.JVM = Void
 type instance MS.XSetupArray CJ.JVM = Void
 type instance MS.XSetupElem CJ.JVM = Void
 type instance MS.XSetupField CJ.JVM = Void